@@ -2,8 +2,8 @@ package com.wangnan.currentactivity.ui.activity;
 
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.SwitchCompat;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.SwitchCompat;
 import android.text.Html;
 import android.view.View;
 import android.widget.LinearLayout;
